Add a profile sidebar panel
